+++ docs/search/all_1.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['attr_0',['attr',['../group__publictypes.html#a5dd5e64c144c5652c8bfea5d9f1eec9b',1,'sir_textstyle']]]
+  ['example_2ec_0',['example.c',['../example_8c.html',1,'']]]
 ];