+++ docs/search/all_5.js
@@ -1,5 +1,4 @@
 var searchData=
 [
-  ['emph_0',['EMPH',['../siransimacros_8h.html#a83167d57009cd2bb4a9116de97039bf3',1,'siransimacros.h']]],
-  ['example_2ec_1',['example.c',['../example_8c.html',1,'']]]
+  ['report_5ferror_0',['report_error',['../example_8c.html#a929f3d3b8335f1827cf00af79a26ade7',1,'example.c']]]
 ];