+++ docs/search/searchdata.js
@@ -1,16 +1,11 @@
 var indexSectionsWithContent =
 {
-  0: "_abcdefgilmnoprstuwy",
+  0: "celmprst",
   1: "st",
-  2: "es",
+  2: "e",
   3: "mrs",
-  4: "_abcdfilmnos",
-  5: "s",
-  6: "s",
-  7: "s",
-  8: "bcdegimrsuwy",
-  9: "dfpt",
-  10: "clt"
+  4: "t",
+  5: "clpt"
 };
 
 var indexSectionNames =
@@ -19,13 +14,8 @@ var indexSectionNames =
   1: "classes",
   2: "files",
   3: "functions",
-  4: "variables",
-  5: "typedefs",
-  6: "enums",
-  7: "enumvalues",
-  8: "defines",
-  9: "groups",
-  10: "pages"
+  4: "groups",
+  5: "pages"
 };
 
 var indexSectionLabels =
@@ -34,12 +24,7 @@ var indexSectionLabels =
   1: "Data Structures",
   2: "Files",
   3: "Functions",
-  4: "Variables",
-  5: "Typedefs",
-  6: "Enumerations",
-  7: "Enumerator",
-  8: "Macros",
-  9: "Modules",
-  10: "Pages"
+  4: "Modules",
+  5: "Pages"
 };